requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-WTF==1.2.2
frozenlist==1.5.0
gql==3.5.0
graphql-core==3.2.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
nltk==3.9.1
numpy==2.1.3
pgzero==1.2.1
propcache==0.2.1
pygame==2.6.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
